requirements.txt
fastapi
uvicorn
torch
transformers
pydantic
numpy
scikit-surprise
pandas